refactor(web): split the release history into current + archive
releases.ts was 8,788 lines / 667 KB, and all 188 KB gzip of it sat in
the initial bundle because App.tsx imported the release-notes page
statically.

The 496 entries are now split verbatim into releaseNotes/current.ts (the
28 since the last epoch boundary - the file every PR edits) and
releaseNotes/archive.ts (the other 468). The About-box copy moves to
appInfo.ts so AboutModal and Help, both eager, no longer sit in the same
module as the archive. The page is a lazy route, so the archive lands in
its chunk instead of the main bundle.

The split was verified by reconstructing the original array from the two
halves and comparing it byte for byte against the committed file.

Note the directory is releaseNotes/, not releases/: .gitignore carries
the Visual Studio build-output rule [Rr]eleases/, which silently ignored
the whole directory - it would have been absent from this commit and
broken CI while looking correct locally.

New union tests replace releases.test.ts and cover the failure the split
introduces: an entry lost down the gap between the halves, or duplicated
across both. Both were mutation-verified.
