+++ apps/web/src/lib/appInfo.ts
@@ -1,0 +1,53 @@
+/// Static facts about the app, and the About-box copy. Deliberately separate from the release
+/// history: `AboutModal` and `Help` are eager, and while they sat in the same module as the
+/// release archive, keeping that archive out of the initial bundle depended on the bundler
+/// tree-shaking it away rather than on the module graph. Now it is structural.
+
+export const TAGLINE = "Smart Meeting Transcription";
+export const GITHUB_URL = "https://github.com/kenhayward/Diariz";
+export const COPYRIGHT = "© Ken Hayward";
+export const LICENSE = "AGPL-3.0";
+
+/// Short summary of what the app does today, shown in the About box (markdown). Update this whenever
+/// the app's scope changes.
+export const CAPABILITIES = `
+Diariz turns your meetings into searchable, speaker-labelled transcripts, then summarises, tags, and lets you chat over them.
+
+| Feature | Description |
+| --- | --- |
+| **Capture** | Record from your mic (device picker, capture tuning, live level meter, pause/resume), system audio, or both mixed together on one device. Pressing **Stop** shows a counting "Uploading... 4s" in the status bar until the recording is safely up. System audio works in Chromium browsers ("Share audio", though on Linux only when sharing a browser tab) and seamlessly in the desktop app; schedule a recording to auto-stop at a set time or after 15/30/60 minutes; or upload files (WAV, MP3, FLAC, Ogg/Oga/Opus, WebM, M4A/M4B, AAC) - or drop a **video** (MP4, M4V, MOV, MKV, WebM, TS/M2TS, 3GP) and its audio is extracted in your browser and uploaded on its own, so the video is never sent or stored. Dropped onto the list they land in the folder you dropped them on, and everything else follows your placement preference. |
+| **Record a calendar meeting** | Join a meeting from your calendar and it records in one click, named after the invite rather than the clock and linked to that meeting from the start (bringing any prep notes with it). Optionally let it end itself - a set number of minutes after the meeting was due to finish, or after a run of silence once everyone has left; if people are still talking when the meeting's scheduled end arrives it asks whether to keep recording instead of cutting you off, doubling the wait each time you say yes, and always tells you why once a recording does end on its own. Joining a second meeting finishes and files the first automatically. It knows the major services by name - Teams (including the short teams.microsoft.com/meet links), Zoom, Google Meet, Webex, Whereby and GoToMeeting - so it opens the link that joins the meeting rather than a help page, a dial-in list or whatever the organiser pasted above them, and finds it even when the invite hides it in the body. |
+| **Recurring meetings** | A calendar event that repeats is marked with a **Repeats** badge, and both the event and a linked recording list your earlier recordings of the same meeting so you can jump straight back. |
+| **Transcribe & diarize** | Server-side WhisperX (word-level timestamps) with pyannote speaker diarization; speaker-labelled, editable, playable segments you can re-transcribe any time. Pin the spoken language on a recording or as your default rather than letting the model guess it, and a language with no word-aligner still produces a transcript. Optionally have every recording's consecutive same-speaker rows merged automatically once it finishes transcribing. A segment shared by two voices can be **split at an exact word** and the new part handed to whoever actually said it. |
+| **Recording hub** | Every meeting opens on a hub: a summary card (meeting type, key facts, the summary inline) over tiles for transcript, actions, speakers, notes, files, and formulas - each showing its count and a preview, with new-note / add-file / run-formula available in place. The transcript embeds a conversation-flow player showing who spoke when, which doubles as the scrubber. |
+| **People and speaker identification** | One shared directory of the people in your meetings, where a voiceprint is optional - someone can be listed with no biometric held, or opt out, which erases theirs. Enrol a voice once and Diariz recognises it across later recordings (SpeechBrain voiceprints); rename, merge, and erase them (biometric data). Assign a speaker from the Speakers tab or straight from the label on any transcript row, as you read or listen. A **People** directory opens over whatever you are reading; it lists everyone one to a line, searchable by name, email or company; a person carries a job title, company, email, phone and an internal/external marker, and likely duplicates are pointed out for you to merge. The Speakers tab shows an identified speaker's title and company inline, with a contact card (email and phone as links) above their segments and a pencil to edit that person in place. Browsing the directory needs the Manage people permission; opting yourself out, and searching to name a speaker, never do. Your own entry is shown read-only on Preferences -> Profile - the name you appear under in transcripts, and whether a voiceprint is held for you. Each person's card has a **Voiceprint** tab listing every recording they appear in - not only the ones enrolled by hand - with how each was attributed, how much they speak in it, and tick boxes to add or drop it from training or to keep only certain segments, re-embedding from exactly what you picked. Play any segment to hear the voice behind the print; audio from a recording you do not own needs the Manage voiceprints permission, which plays only what that person said and logs every use. Where two people share a name, every list shows which Diariz account each one is - or says plainly that there is none - so a merge is decidable. When a voice is close but not certain, Diariz asks rather than guessing or staying silent - on the speaker in the transcript, and gathered under **Review Voice Matches**, a two-panel window listing the waiting voices beside what each one said, a line per segment with a play/stop button and a control to take it out, plus a Play all that runs through the kept segments highlighting each as it goes; excluding a segment keeps it out of the voiceprint, and voices whose audio has been deleted are not offered at all, since there is nothing left to listen to; a Platform Administrator sets how close is close enough, and can re-scan earlier recordings against the current settings. Each recording on that tab carries its own verdict - matches the others, a different recording condition, resembles nothing at all, or **sounds more like somebody else**, naming who - with the ones worth a listen sorted to the top and a tick box to show only those. Confirm a recording once you have listened and it leaves the review queue. Figures are stated as similarity, so a high number means a good match. Where a recording turns out to be somebody else, reassign it - or create that person, mark it as overlapping speech, or unlink it - from the same row, for recordings you own. In the directory, a person with a shaky voiceprint or a likely duplicate carries a short warning on their own row rather than a panel over the list, and a **Needs review** filter narrows to them. |
+| **Notes** | Take your own note lines live during a meeting (timestamped, crash-safe); they appear inline in the transcript at the moment you wrote them, steer the minutes, and can be woven into an enhanced-notes section linking to the exact transcript moments. In the desktop app, pop the notes out into a small always-on-top window that floats over a full-screen call, so a single monitor is no longer a reason to lose sight of the meeting. |
+| **Meeting screenshots** | Capture the screen during a recording from the desktop app - a hotkey, the tray menu, or the app itself - choosing a screen or a rectangle on the first capture and reusing it after; or turn on **Auto-capture** and it takes one every time the screen settles on something new, capturing a presentation slide by slide on its own. Captures appear in the transcript at the moment they were taken, as a full-size viewer with zoom and pan to read a capture at native resolution, and in a Notes-tab section. From either the viewer's own **Add to chat context** button or by **dragging the thumbnail into the chat prompt**, a capture goes to a vision-capable model to ask about. Where an administrator has routed an **OCR** model, two more buttons read the text off a capture - into the chat prompt, or saved as a Markdown attachment on the meeting - cached so the second costs nothing, tables converted to Markdown so they render as tables, and always stamped with the model that read it and marked unverified. |
+| **Summaries & minutes** | Auto summary plus full professional meeting minutes, editable in a rich editor and emailable. A meeting type carries no prompts of its own: it names the **formula** that generates its minutes, plus any others to run alongside (their documents appear in the Formulas tab). Minutes and formulas are the same machinery - any formula you can use can produce your minutes. Templates are built from blocks (H1-H3 headings, literal text, substituted details, model prompts, rules, drag-to-reorder) with JSON import/export. |
+| **Action items** | Auto-extracted with owner and deadline, tracked across every meeting with completion and a person filter, linking back to the transcript. |
+| **Tag cloud** | Tags are yours to add - a Tags pill on each meeting suggests topics automatically extracted from it, and adopting one builds the weighted cloud on the Tags tab, which shows the meetings behind each tag. |
+| **Chat over transcripts** | Stream answers over one meeting, a folder (its summary/minutes/actions), several selected, or all meetings - context inferred from what you're viewing - with attachments and saved conversations. Where more than one model is offered, pick which one answers - the list names each model, describes what it is for, and marks which can use tools or read images - and change it part-way through a conversation; the whole conversation goes across to the new model. Send in a meeting screenshot - from the capture's own viewer or by dragging its thumbnail - and a vision-capable model will read it. |
+| **Formulas** | Build a document from headings, literal text, substituted meeting details and instructions to the model - with \`$USERNAME\` in a prompt becoming the name the person running it appears under on their transcripts; choose what it may see (transcript, notes, summary, minutes, actions) and run it over a recording - or a whole folder and its sub-folders - to generate a Markdown document you can edit, download, or email. Runs in the background (\"Generating...\" then fills in); re-running one replaces its previous document, and a document you edited by hand is left alone. Personal, platform-wide, or built-in; from the Formulas tab (on a recording or a folder), \`/formula\` in chat, or Claude via MCP; share a personal one so others can find and add it (a live link); admins manage the shared ones. |
+| **Search** | A search box above the meetings list, scoped by default to the folder you are browsing: typing takes the list over, each hit shows the matching words in context plus the folder it lives in, and clicking one opens the transcript at that moment. Matching folders show up too. **Search everywhere** widens it to every room you can see, grouping the results by folder with Folder / Date / Speaker chips to narrow them. Keyword search across your library, upgraded to semantic (meaning-based) search when an embeddings endpoint is configured. |
+| **Chat tools** | The assistant searches your library with built-in tools (who-said-what, attendees, talk time, summaries, email-to-self) and links to the exact segment. Turn the set on or off, and pick tool by tool, on the **Assistant** tab in Preferences. |
+| **Voice dictation** | Speak your chat questions - transcribed into the chat box in Chrome/Edge or via a server speech-to-text endpoint. |
+| **MCP access** | Let an AI assistant read your own meetings over MCP - via OAuth (the claude.ai connector) or a personal token (Claude Desktop, Claude Code), when the platform's MCP toggle is on - including a \`run_formula\` tool to run your saved Formulas. Set up on the **Integrations** tab in Preferences, alongside API access and Automations, a card each. |
+| **Translate** | Translate a whole transcript or a single segment, stored as revisions you can flip back. |
+| **Attachments** | Attach files or URLs (PDF, Office, email, calendar, images) to a recording or directly to a folder, edit Markdown attachments in place, save a chat conversation as an attachment with /attach, and optionally feed them to chat - either every attachment in scope with one tick box, or a single one dragged straight onto the chat box by its handle and read back by clicking the pill it leaves. |
+| **Rooms** | A private Personal Room per account plus shareable Rooms: invite users and groups with per-member permissions. Each Shared Room has its **own folder structure** (folders/sub-folders, drag-and-drop, per-room order) and its own List/Calendar/Actions/Tags scoped to it; record or upload files straight into a room (your Personal Room keeps the original), and search + chat over every room you belong to. A member who can read a shared recording sees its notes and screenshots too - only the owner can add, edit, or delete them. Your Google Calendar and its linking stay personal. The switcher shows each room's folder and meeting counts (shared ones labelled), ticks the one you are in, and remembers where you were. Manage rooms from the switcher. |
+| **Organise & merge** | Folders nested up to 8 levels deep with drag-and-drop (dragging one of several ticked rows moves the whole selection); the meetings list drills in one folder at a time (coloured folder rows with counts, a breadcrumb showing the full path with every part clickable, browser back pops a level) so it stays readable however many recordings you have, with a page button beside the breadcrumb's menu opening the folder's own page, a separate target from browsing deeper; an open recording shows its folder path as clickable chips under its name, each taking the list straight to that folder; choose where a new recording is filed (Ungrouped, the open folder, or a specific folder); sort the list by date/time, name or duration (ascending or descending) or keep your manual order, remembered between visits, with each row showing when it was made; cut one or more recordings, or a folder, and paste them into another folder in one move (landing at the bottom, in the order you cut them), with the Paste control showing why it's disabled when a move isn't allowed yet; browse as a list, calendar, actions, or tag cloud; merge recordings into one. |
+| **Folder pages** | Open any folder as a page: a roll-up LLM summary and consolidated minutes across it and its sub-folders, plus every action, note, and attachment aggregated with the meeting each came from. |
+| **Google & calendars** | Optional Google sign-in and read-only Calendar linking, plus subscriptions to public iCalendar (.ics) feeds and - synced by the Windows desktop app on launch, from the tray or on demand - a mirror of your classic Outlook calendar (opt-in, per machine, erasable). All three are set up in one **Calendars** tab in Preferences, a card per source showing its account, its calendars and whether it is live. Every calendar you connect is treated the same: meetings from any of them show on the Calendar tab, match to a recording, and open with their full invite details. A month grid highlights the days that have something on them; picking one lays that day out on an hour axis, with meetings and recordings placed and sized by when they ran, clashes side by side, all-day entries in their own strip, and a line marking the current time. A **Go to today** button on the panel toolbar brings both the day and the month grid back to today after you have browsed ahead. Two toolbar buttons refresh it: **Sync calendar** covers every source you have connected in one go, and **Sync today** does the same for the current day only - seconds rather than the half a minute a full Outlook read takes - with the status bar counting up while either runs. |
+| **LLM usage log** | Every model call recorded - who it was for, which model answered, how long it took, how many tokens, and why the model stopped. A **Cut off** badge flags a reply that hit its token cap, which otherwise looks identical to a model that answered nothing. Operations, Calls and Summary views with filters and a totals row over the whole filter. |
+| **AI models** | A Platform Administrator configures every model the platform calls - its endpoint, key, context window and a one-line description people see when they pick it - and how it is called: temperature, top P, top K, penalties, token caps, reasoning effort, timeout, and whether it supports tool calling or images. Routing is a grid - models down the side, call types across the top, one click to move a call type to another model, a **No model** row for the ones that follow the default, and an **In chat** column marking which models people may pick between for chat. **Add all** adds every chat model on a server in one pass, leaving out embeddings and speech models and flagging any whose context window the server did not report. Parameters live in a drawer with a tab per call type, each parameter inherited, set, or omitted from the request entirely, beside a live preview of the exact request body that call type would send. **Run test** makes a real call with whatever is on screen and reports first-token time, duration, tokens per second and the model's reply - offering, on a failure, the single change that would fix it, and on any result a cURL command, the raw JSON, or a link into the usage log. For Tags, Actions and Summaries the test runs the **real prompt against one of your own recordings** (picked in the panel and remembered), and shows the parsed result the platform would have stored - weighted tags, an action table, a summary and its suggested name - saying so plainly when a reply would have yielded nothing. A whole set can be copied between models. |
+| **API access** | Generate a personal API token (when enabled), read-only or read-write, with an optional expiry date, to call the REST API as yourself; the list shows which tokens are read-only and when each expires. A built-in API reference documents every endpoint - what it does, who may call it, and what it changes. |
+| **Automations (webhooks)** | When enabled, register outbound webhooks from Preferences - a list of what you have, with a composer for making or editing one - that fire when a recording is created, finishes or fails transcription, a summary / meeting minutes / action items / tags become ready (each carrying its output), or a formula finishes or fails. Every recording event also carries an **attendees** list - who spoke, the person they are, their title, company, and internal or external - with email addresses and phone numbers included only when that automation opts in. Signed deliveries with automatic retries (rate-limited per automation and 429-aware), a test-event button, auto-pause after repeated failures, and a Pause / Resume button that stops deliveries reversibly (deleting an automation discards its signing secret; pausing keeps it). Admins can also define named Workflow Signals and wire a platform automation to one, so a formula author just picks "When this finishes, trigger: ..." in the formula editor - no URL or per-user setup - and the formula's output is delivered inline to everyone routed through that signal. |
+| **n8n** | A published community node (\`n8n-nodes-diariz\`): a trigger that registers its own automation and verifies every signed delivery, with a Platform scope for administrators that adds events across every user including Feedback Received, plus an action node covering the whole REST API - with dropdowns listing your real recordings and formulas, files in and out, and a Run Formula step that waits for the document to finish. |
+| **Multi-user & groups** | A **Users & access** console: search and filter every account, and see on one panel what a person's groups actually let them do, alongside their quota, status and storage. Groups grant the five platform permissions - each explained in a sentence rather than a column heading - and carry a name, description and colour. Access requests are a tab with a waiting count, then an approval lifecycle. Per-user isolation and Light/Dark/Auto themes. |
+| **Admin controls** | Storage quotas, optional audio auto-deletion, a default AI request timeout that users can override for their own account, LLM usage logging (on/off, a retention window in days, a streaming-token-count toggle, and a usage viewer to filter, total, and delete the captured log), separate platform toggles for API access, Claude/MCP, and Automations (webhooks), and whole-platform backup & restore. |
+| **Provide Feedback** | Any signed-in user can describe something that looks or behaves wrong from the account menu; a scrubbed technical trail of recent app activity is attached automatically. Readable and deletable only by a Platform Administrator, in a Feedback tab in Settings, and can raise a \`feedback.submitted\` automation event - the submitter's words reach it only when that automation ticks Include What The Person Wrote. |
+| **Install as an app** | Install Diariz from a Chromium browser and it gets a launcher entry, an icon, and its own window with no browser chrome - the account menu offers it whenever your browser can. This is how Linux gets an app-like Diariz, since the desktop installer is Windows and macOS only. |
+| **Help & documentation** | A browsable help section at \`/help\` with a grouped article tree and instant search, plus contextual help throughout the app: a small \`?\` next to a feature opens a short explanation with a link straight to the full article. An **Advanced and admin** section covers configuring formulas, meeting types, Workflow Signals and webhooks, users and permissions, connecting Claude over MCP, building n8n and Zapier workflows, and the REST API. A **CRM integration** section covers wiring Diariz to any CRM through n8n or Zapier, with a worked EspoCRM example. |
+
+The interface is localized (English, Spanish, French, German), and languages are community-extensible via simple JSON files.
+`.trim();
